requirements.txt
streamlit
requests
transformers
datasets
torch
python-dotenv
wikipedia
